requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# LLM & AI
groq==0.4.1
openai==1.10.0

# Image Processing
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring & Logging
python-json-logger==2.0.7

# Optional: Production Server
gunicorn==21.2.0

# Optional: Task Queue (if implementing async tasks)
celery==5.3.4
redis==5.0.1

# Optional: WebSocket support
websockets==12.0

# Optional: Email support
aiosmtplib==3.0.1

# Optional: Rate Limiting (advanced)
slowapi==0.1.9

# Optional: CORS (already in FastAPI)
# fastapi already includes CORS middleware
